+++ Source/Krypton Components/Krypton.Toolkit/Controls Toolkit/KryptonCustomPaletteBase.cs
@@ -5637,6 +5637,7 @@ namespace Krypton.Toolkit
             {
                 case PaletteState.Disabled:
                     return header.StateDisabled;
+                case PaletteState.Tracking: // #1729 KHeader does not implement the tracking state, default to normal
                 case PaletteState.Normal:
                     return header.StateNormal;
                 default:
@@ -5671,7 +5672,7 @@ namespace Krypton.Toolkit
                 case PaletteState.Disabled:
                     return label.StateDisabled;
                 case PaletteState.Normal:
-                case PaletteState.ContextNormal:    // Occurrs from the TreeGrid
+                case PaletteState.ContextNormal:    // Occurs from the TreeGrid
                 case PaletteState.Tracking:
                 case PaletteState.Pressed:
                     return label.StateNormal;
@@ -5734,6 +5735,8 @@ namespace Krypton.Toolkit
                     return ContextMenu.StateNormal.ItemImage.Back;
                 case PaletteState.CheckedNormal:
                     return ContextMenu.StateChecked.ItemImage.Back;
+                case PaletteState.CheckedTracking:
+                    return ContextMenu.StateHighlight.ItemHighlight.Back;
                 default:
                     // Should never happen!
                     Debug.Assert(false);
@@ -5785,6 +5788,8 @@ namespace Krypton.Toolkit
                     return ContextMenu.StateNormal.ItemImage.Border;
                 case PaletteState.CheckedNormal:
                     return ContextMenu.StateChecked.ItemImage.Border;
+                case PaletteState.CheckedTracking:
+                    return ContextMenu.StateHighlight.ItemHighlight.Border;
                 default:
                     // Should never happen!
                     Debug.Assert(false);